[awu19] update the semantic visitor to check for integer overflow exceptions with immediates

+++ src/main/kotlin/wacc_05/front_end/SemanticErrors.kt
@@ -40,5 +40,5 @@ interface SemanticErrors {
 
     fun invalidExitType(ctx: ParserRuleContext, actual: TypeIdentifier)
 
-    fun integerOverflow(ctx: ParserRuleContext, value: Int)
+    fun integerOverflow(ctx: ParserRuleContext, value: Long)
 }
